Final pass for speedups. ~98ms runtime.

+++ Problem045.cs
@@ -7,7 +7,7 @@ public class Problem045 : Problem {
 
     private long FindNextTphNumber(int n) {
         n++; //since n is valid, and we want to find the next one.
-        int i = n;
+        long i = n;
         while (true) {
             long hex = i * (2L * i - 1);
             if (Library.IsPentagon(hex)) return hex;
